Moves nameToInitials to Avatar
