+++ src/components/Avatar/Avatar.jsx
@@ -3,13 +3,12 @@ import PropTypes from 'prop-types'
 import getValidProps from '@helpscout/react-utils/dist/getValidProps'
 import classNames from 'classnames'
 import { getEasingTiming } from '../../utilities/easing'
-import { nameToInitials } from '../../utilities/strings'
 import Icon from '../Icon'
 import StatusDot from '../StatusDot'
 import Tooltip from '../Tooltip'
 import AvatarCrop from './Avatar.Crop'
 import AvatarImage from './Avatar.Image'
-import { getImageSrc } from './Avatar.utils'
+import { getImageSrc, nameToInitials } from './Avatar.utils'
 import { AvatarListContext } from '../AvatarList/AvatarList'
 import {
   ActionUI,
