+++ src/components/Avatar/Avatar.utils.js
@@ -21,3 +21,24 @@ export const getAnimationProps = props => {
     animation,
   }
 }
+
+export const nameToInitials = (name = '') => {
+  // Returning early if undefined to avoid casting undefined to "undefined"
+  if (!name) {
+    return ''
+  }
+
+  // Trim trailing whitespace
+  name = (name + '').trim()
+  if (!name.length) {
+    return ''
+  }
+
+  const words = name
+    .split(' ')
+    .filter(w => w !== '')
+    .map(w => w[0])
+    .map(w => w.toUpperCase())
+
+  return words.length === 1 ? words[0] : words[0] + words[words.length - 1]
+}
