+++ src/utilities/strings.js
@@ -1,27 +1,6 @@
 import isString from 'lodash.isstring'
 import { normalizeUrl } from './urls'
 
-export const nameToInitials = (name = '') => {
-  // Returning early if undefined to avoid casting undefined to "undefined"
-  if (!name) {
-    return ''
-  }
-
-  // Trim trailing whitespace
-  name = (name + '').trim()
-  if (!name.length) {
-    return ''
-  }
-
-  const words = name
-    .split(' ')
-    .filter(w => w !== '')
-    .map(w => w[0])
-    .map(w => w.toUpperCase())
-
-  return words.length === 1 ? words[0] : words[0] + words[words.length - 1]
-}
-
 export const isWordString = word => {
   return !!(isString(word) && word.length)
 }
